requirements.txt
requests
python-dotenv
pandas
supabase
rich
tqdm


